fix:レビュー指摘箇所修正
