+++ apps/backend/drizzle/0004_cold_prowler.sql
@@ -3,7 +3,7 @@ CREATE TABLE "competitions" (
 	"name" varchar(100) NOT NULL,
 	"start_date" timestamp with time zone NOT NULL,
 	"end_date" timestamp with time zone NOT NULL,
-	"is_active" boolean DEFAULT false NOT NULL,
+	"is_active" boolean DEFAULT true NOT NULL,
 	"created_at" timestamp with time zone DEFAULT now(),
 	"updated_at" timestamp with time zone DEFAULT now()
 );
