+++ apps/backend/drizzle/meta/0004_snapshot.json
@@ -116,7 +116,7 @@
           "type": "boolean",
           "primaryKey": false,
           "notNull": true,
-          "default": false
+          "default": true
         },
         "created_at": {
           "name": "created_at",
