+++ packages/shared/src/types.ts
@@ -22,6 +22,27 @@ export interface CreateBodyRecordRequest {
 
 export interface UpdateBodyRecordRequest extends CreateBodyRecordRequest {}
 
+export interface RankingParticipant {
+  rank: number;
+  userId: string;
+  username: string;
+  baselineWeight: number;
+  currentWeight: number;
+  weightLossRate: number; // %
+  baselineBodyFat: number;
+  currentBodyFat: number;
+  bodyFatLossRate: number; // %
+  totalScore: number; // %
+  recordedAt: string;
+}
+
+export interface RankingData {
+  competitionName: string;
+  startDate: string;
+  endDate: string; // "YYYY-MM-DD" or ""
+  rankings: RankingParticipant[];
+}
+
 export interface ApiResponse<T> {
   data?: T;
   error?: string;
